feat: implement filters and sorting query managment

+++ Controllers/CommentController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using start.Dtos.Comment;
 using start.Interfaces;
 using start.Mappers;
@@ -19,23 +20,29 @@ namespace start.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllAsync();            
 
             return Ok(comments.Select(comment => comment.ToCommentDto()));
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById(int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepository.GetByIdAsync(id);
             if (comment == null) return NotFound();
 
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
          public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto createCommentDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             if (!await _stockRepository.StockExistsAsync(stockId)) return BadRequest("Stock does not exist.");
 
             var comment = createCommentDto.ToCommentFromCreateDto(stockId);
@@ -44,18 +51,22 @@ namespace start.Controllers
             return CreatedAtAction(nameof(GetById), new { id = comment.Id }, comment.ToCommentDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id}:int")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateCommentDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepository.UpdateAsync(id, updateCommentDto.ToCommentFromUpdateDto());
             if (comment == null) return NotFound(); 
 
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id}:int")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepository.DeleteAsync(id);
             if (comment == null) return NotFound();
